massive overhaul to integrade the pc version of dave
