--- WEB/login/login.js
@@ -1,30 +1,0 @@
-document.addEventListener('DOMContentLoaded', createStars);
-
-// ----- Stars Background -----
-
-function createStars() {
-    const starsContainer = document.getElementById('stars');
-    if (!starsContainer) return;
-    
-    const starCount = 200;
-    
-    for (let i = 0; i < starCount; i++) {
-        const star = document.createElement('div');
-        star.className = 'star';
-        
-        const size = Math.random() * 3 + 1;
-        const x = Math.random() * 100;
-        const y = Math.random() * 100;
-        const duration = Math.random() * 3 + 2;
-        
-        star.style.cssText = `
-            left: ${x}%;
-            top: ${y}%;
-            width: ${size}px;
-            height: ${size}px;
-            animation-duration: ${duration}s;
-        `;
-        
-        starsContainer.appendChild(star);
-    }
-}
