+++ public/admin-panel/admin-panel.css
