+++ public/admin-panel/admin-panel.js
