+++ public/home-page/home-page.html
@@ -1,0 +1,92 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>PRISM - Home</title>
+    <link rel="icon" type="image/png" href="/images/prism.favicon.png">
+    <link rel="stylesheet" href="/css/home-page.css">
+</head>
+<body>
+    <div class="stars-container" id="stars"></div>
+
+    <div class="content">
+        <nav class="navbar">
+            <div class="logo">
+                <img src="/images/prism.logo.png" alt="PRISM Logo" class="img-logo">
+            </div>
+            <button class="hamburger" id="hamburgerBtn" type="button" aria-label="Toggle navigation menu">
+                <img src="/images/menu.png" alt="Menu">
+            </button>
+            <ul class="nav-links" id="navLinks">
+                <li><a href="/home-page/home-page.html"><span>HOME</span></a></li>
+                <li><a href="/product-scanner/product.scanner.html"><span>PRODUCT SCANNER</span></a></li>
+                <li><a href="/dashboard/dashboard.html"><span>DASHBOARD</span></a></li>
+                <li><a href="/marketplace/marketplace.html"><span>MARKETPLACE</span></a></li>
+                <li><a href="/admin-panel/admin-panel.html"><span>ADMIN PANEL</span></a></li>
+            </ul>
+        </nav>
+
+        <section class="home-page">
+            <div class="home-content">
+                <div class="home-text">
+                    <p>Price Intelligence and Smart Monitoring</p>
+                    <h1>Protect Every Peso with <span>PRISM</span></h1>
+                    <p class="home-description">
+                        Monitor market prices, detect suspicious spikes, and connect with trusted suppliers in one platform.
+                    </p>
+                    <div class="home-buttons">
+                        <button class="login" type="button" onclick="window.location.href='/login/login.html'">Log In</button>
+                        <button class="create-account" type="button" onclick="window.location.href='/create-account/create-account.html'">Create Account</button>
+                    </div>
+                </div>
+                <div class="home-image">
+                    <img src="/images/home.pages.png" alt="PRISM market insights dashboard">
+                </div>
+            </div>
+        </section>
+
+        <section class="market-impact">
+            <h2 class="section-title">Live Impact Overview</h2>
+            <div class="impact-grid">
+                <article class="impact-card">
+                    <div class="impact-number" id="shopsCount">0</div>
+                    <div class="impact-label">Active Partner Shops</div>
+                    <div class="impact-trend positive">Growing network</div>
+                </article>
+                <article class="impact-card">
+                    <div class="impact-number" id="fairnessCount">0</div>
+                    <div class="impact-label">Fair Price Match Rate</div>
+                    <div class="impact-trend positive">Improving weekly</div>
+                </article>
+                <article class="impact-card">
+                    <div class="impact-number" id="anomaliesCount">0</div>
+                    <div class="impact-label">Detected Price Anomalies</div>
+                    <div class="impact-trend warning">Under review</div>
+                </article>
+            </div>
+        </section>
+
+        <footer class="prism-section footer">
+            <div class="prism-container footer-container">
+                <div class="footer-left">
+                    <span>&copy; PRISM 2026</span>
+                    <span>&bull;</span>
+                    <a href="/terms/terms.html">Terms</a>
+                    <span>&bull;</span>
+                    <a href="/privacy/privacy.html">Privacy</a>
+                    <span>&bull;</span>
+                    <a href="/faqs/faqs.html">FAQs</a>
+                    <span>&bull;</span>
+                    <a href="/about/about.html">About</a>
+                </div>
+                <div class="footer-right">
+                    <pre>Transparency    |    Protection    |    Integrity    |    Intelligence</pre>
+                </div>
+            </div>
+        </footer>
+    </div>
+
+    <script src="/js/home-page.js"></script>
+</body>
+</html>
