+++ public/login/login.html
@@ -1,0 +1,70 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=5.0, user-scalable=yes">
+    <title>PRISM – Login</title>
+    <link rel="icon" type="image/png" href="/images/prism.favicon.png">
+
+<!-- CSS -->
+
+    <link rel="stylesheet" href="login.css">
+
+</head>
+
+<body>
+
+<!-- Stars Background -->
+
+  <div class="stars-container" id="stars"></div>
+
+<!-- Login panel -->
+
+  <div class="login-box">
+
+    <h2>Log in to Your Account</h2>
+
+    <button class="social-btn google">
+      <img src="/images/google.png" alt="Google">
+      Continue with Google
+    </button>
+
+    <button class="social-btn facebook">
+      <img src="/images/facebook.png" alt="Facebook">
+      Continue with Facebook
+    </button>
+
+    <div class="divider"></div>
+
+    <input type="email" placeholder="Email address">
+    <input type="password" placeholder="Password">
+
+    <button class="login-btn">LOGIN</button>
+
+    <p class="signup">
+      Don’t have an account?
+      <a href="/create-account/create-account.html">Sign up</a>
+    </p>
+  </div>
+
+  <footer>
+    <ul class="footer-links">
+        <p>© prism.ph</p>
+        <p> • </p>
+        <li> <a href="/terms/terms.html"> Terms </a> </li>
+        <p> • </p>
+        <li> <a href="/privacy/privacy.html"> Privacy</a> </li>
+        <p> • </p>
+        <li> <a href="/faqs/faqs.html"> FAQs </a> </li>
+        <p> • </p>
+        <li> <a href="/about/about.html"> About</a> </li>
+    </ul>
+</footer>
+
+<!-- JavaScript  -->
+
+    <script src="login.js"></script>
+
+</body>
+
+</html>
