+++ public/mobile/m-dashboard.html
@@ -1,30 +1,30 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <link rel="icon" type="image/png" href="images/prism.favicon.png">
+    <link rel="icon" type="image/png" href="/images/prism.favicon.png">
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>PRISM – Dashboard</title>
     <script src="https://cdn.tailwindcss.com"></script>
-    <link rel="stylesheet" href="/public/css/main..css">
+    <link rel="stylesheet" href="/css/main..css">
     <style>
-      @font-face { font-family: 'Neue Montreal'; src: url(fonts/ppneuemontreal-medium.ttf); }
-      @font-face { font-family: 'Neue Book'; src: url(fonts/ppneuemontreal-book.ttf); }
+      @font-face { font-family: 'Neue Montreal'; src: url(/fonts/ppneuemontreal-medium.ttf); }
+      @font-face { font-family: 'Neue Book'; src: url(/fonts/ppneuemontreal-book.ttf); }
     </style>
 </head>
 <body class="bg-black text-white min-h-screen flex flex-col">
   <div class="w-full flex items-center justify-between px-[25px] py-4 bg-black text-white">
-    <div class="logo"><div class="icon flex items-center gap-3 shrink-0"><img src="images/prism.logo.wtext.png" alt="PRISM logo" class="w-10 h-10 object-contain"></div></div>
+    <div class="logo"><div class="icon flex items-center gap-3 shrink-0"><img src="/images/prism.logo.wtext.png" alt="PRISM logo" class="w-10 h-10 object-contain"></div></div>
     <div class="flex-1 flex justify-center text-justify" style="font-family: 'Neue Montreal';">
       <ul class="flex items-center gap-12 text-sm tracking-wide">
-        <li><a href="/index.html" class="flex items-center gap-2 hover:text-green-400 transition"><img src="images/home.svg" alt="Home" class="w-4 h-4"><span>HOME</span></a></li>
-        <li><a href="/product-scanner.html" class="flex items-center gap-2 hover:text-green-400 transition"><img src="images/product scanner_white.png" alt="Product Scanner" class="w-4 h-4"><span>PRODUCT SCANNER</span></a></li>
-        <li><a href="/dashboard.html" class="flex items-center gap-2 hover:text-green-400 transition"><img src="images/dashboard.svg" alt="Dashboard" class="w-4 h-4"><span>DASHBOARD</span></a></li>
-        <li><a href="/marketplace.html" class="flex items-center gap-2 hover:text-green-400 transition"><img src="images/marketplace.svg" alt="MarketPlace" class="w-4 h-4"><span>MARKET PLACE</span></a></li>
-        <li><a href="/admin-panel.html" class="flex items-center gap-2 hover:text-green-400 transition"><img src="images/admin.panel.svg" alt="Admin Panel" class="w-4 h-4"><span>ADMIN PANEL</span></a></li>
+        <li><a href="/home-page/home-page.html" class="flex items-center gap-2 hover:text-green-400 transition"><img src="/images/home.svg" alt="Home" class="w-4 h-4"><span>HOME</span></a></li>
+        <li><a href="/product-scanner/product.scanner.html" class="flex items-center gap-2 hover:text-green-400 transition"><img src="/images/product scanner_white.png" alt="Product Scanner" class="w-4 h-4"><span>PRODUCT SCANNER</span></a></li>
+        <li><a href="/dashboard/dashboard.html" class="flex items-center gap-2 hover:text-green-400 transition"><img src="/images/dashboard.svg" alt="Dashboard" class="w-4 h-4"><span>DASHBOARD</span></a></li>
+        <li><a href="/marketplace/marketplace.html" class="flex items-center gap-2 hover:text-green-400 transition"><img src="/images/marketplace.svg" alt="MarketPlace" class="w-4 h-4"><span>MARKET PLACE</span></a></li>
+        <li><a href="/admin-panel/admin-panel.html" class="flex items-center gap-2 hover:text-green-400 transition"><img src="/images/admin.panel.svg" alt="Admin Panel" class="w-4 h-4"><span>ADMIN PANEL</span></a></li>
       </ul>
     </div>
-    <div><img src="images/menu.png" alt="Menu" class="w-12 h-12 cursor-pointer"></div>
+    <div><img src="/images/menu.png" alt="Menu" class="w-12 h-12 cursor-pointer"></div>
   </div>
   <main class="flex-1 flex items-center justify-center px-6">
     <div class="text-center max-w-3xl">
