+++ tsconfig.json
@@ -22,4 +22,4 @@
   ]
 }
 
-// 
+//
